[FLINK-24325][tests] Update Elasticsearch CI tests
