+++ flink-end-to-end-tests/flink-quickstart-test/pom.xml
@@ -56,7 +56,7 @@ under the License.
 		</dependency>
 		<dependency>
 			<groupId>org.apache.flink</groupId>
-			<artifactId>flink-connector-elasticsearch5</artifactId>
+			<artifactId>flink-connector-elasticsearch7</artifactId>
 			<version>${project.version}</version>
 		</dependency>
 	</dependencies>
