+++ tools/ci/java-ci-tools/src/main/resources/modules-skipping-deployment.modulelist
@@ -36,7 +36,6 @@ flink-streaming-kinesis-test
 flink-tpch-test
 flink-streaming-kafka-test-base
 flink-heavy-deployment-stress-test
-flink-elasticsearch5-test
 flink-high-parallelism-iterations-test
 flink-end-to-end-tests-common-kafka
 flink-end-to-end-tests-pulsar
